+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* Add `has` function
+
 ## 3.1.2 (2019-04-14)
 
 * Fix issue in `union` function when handling arguments with duplicate values